rename resource pkg.
